+++ order-service/src/main/java/com/ecommerce/order/dto/ProductResponse.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record ProductResponse(
         Long id,
         String name,
-        BigDecimal price,
+        Long price,
         Integer stockQuantity
 ) {
 }
